Move TDDiskState to separate file
