+++ ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/dirty_map/dirty_map.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "ddisk_state.h"
 #include "inflight_info.h"
 #include "range_locker.h"
 
@@ -155,84 +156,6 @@ struct TSyncHint
 };
 
 ////////////////////////////////////////////////////////////////////////////////
-
-class TDDiskState
-{
-public:
-    enum class EState
-    {
-        Disabled,   // There are no DDisks with data on the host and DDisk
-                    // cannot be used.
-
-        Operational,   // The DDisk is fully functional and can be read from
-                       // anywhere. BehindField and AheadField are empty.
-
-        Fresh,   // The ddisk is only partially filled, and you can only read
-                 // from the blocks below the OperationalBlockCount.
-                 // The AheadField shows which ranges flushed over watermark and
-                 // can be read from. The BehindField shows which ranges
-                 // outdated and can't be read.
-    };
-
-    enum class EFlushCompletion
-    {
-        Completed,   // Data flushed to DDisk
-        Missed,      // Data not flushed to DDisk
-    };
-
-    // Enables the use of DDisk. If the operational blocks count less then total
-    // block count, then the DDisk is only partially filled (fresh).
-    void Init(ui64 totalBlockCount, ui64 operationalBlockCount);
-
-    // Completely disables DDisk usage.
-    void SwitchOffline();
-
-    [[nodiscard]] bool IsLagging() const;
-    // DDisk has stopped receiving writes. Now the written ranges are
-    // interpreted as "bad" and added to the BehindField.
-    void StartLagging();
-    // DDisk now receive all writes. The written ranges are interpreted as
-    // "good" and removed from the BehindField.
-    void StopLagging();
-    // Is it necessary to receive information about all written ranges. If true
-    // is returned, it means that all ranged that have been flushed must be
-    // passed to the OnRangeFlushed() method.
-    [[nodiscard]] bool IsTrackingEnabled() const;
-    // Updates the BehindField and the Ahead Field if required.
-    void OnRangeFlushed(TBlockRange64 range, EFlushCompletion flush);
-
-    [[nodiscard]] EState GetState() const;
-    [[nodiscard]] bool CanReadFromDDisk(TBlockRange64 range) const;
-
-    [[nodiscard]] std::optional<TBlockRange64> GetFreshRange() const;
-    void RangeSynced(TBlockRange64 range);
-
-    void UpdateWatermarkDebugOnly(ui64 blockCount);
-    [[nodiscard]] TString DebugPrint() const;
-    [[nodiscard]] TString DebugPrintAhead() const;
-    [[nodiscard]] TString DebugPrintBehind() const;
-
-private:
-    [[nodiscard]] bool IsFresh() const;
-    void UpdateState(bool force);
-    void AddAhead(TBlockRange64 range);
-
-    EState State = EState::Disabled;
-
-    ui64 TotalBlockCount = 0;
-
-    // If the block address below OperationalBlockCount, then it can be read
-    // from DDisk (except BehindField).
-    ui64 OperationalBlockCount = 0;
-
-    // Lagging means that flush operations are not performed and DDisk has
-    // outdated data in the ranges listed in the BehindField.
-    bool Lagging = false;
-    TBlockRangeField BehindField;
-    // When a user writes to a range above OperationalBlockCount, this range has
-    // up-to-date data and does not require sync.
-    TBlockRangeField AheadField;
-};
 
 struct TPBufferCounters
 {
