+++ ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/dirty_map/ut/ya.make
@@ -1,6 +1,7 @@
 UNITTEST_FOR(ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/dirty_map)
 
 SRCS(
+    ddisk_state_ut.cpp
     dirty_map_ut.cpp
     inflight_info_ut.cpp
     range_locker_ut.cpp
